Patch more things to speed up pillowtop tests
